pyproject.toml
[tool.poetry]
name = "matgame"
version = "0.1.0"
description = "matific demo gameapp"
authors = ["self <geass.of.code@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.15"
PyGithub = "^1.55"
asyncio = "^3.4.3"
datatest = "^0.11.1"
django-extensions = "^3.1.5"
make = "^0.1.6"
openpyxl = "^3.0.10"
pandas = "^1.5.3"
numpy = "^1.26.4"
pytest = "^8.3.2"
python-dotenv = "^0.20.0"
python-multipart = "^0.0.7"
requests = "^2.28.1"
snaql = "0.4.2"
tabulate = "^0.8.9"
uvicorn = "^0.17.6"
pytest-django = "^4.5.2"
psycopg2 = "^2.9.5"
gunicorn = "^22.0.0"
whitenoise = "^6.2.0"
django-prometheus = "^2.2.0"
pyarrow = "^14.0.1"
factory-boy = "^3.2.1"
flaky = "^3.8.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.26.2"
python-dateutil = "^2.8.2"
django-filter = "^23.2"
django-hijack = "^3.4.1"
coverage = "^7.3.1"

moto = "^4.2.5"
pytest-xdist = "^3.5.0"
pytest-cov = "^4.1.0"
aiohttp = "^3.9.5"
pydantic = "^2.8.2"
django-cors-headers = "^4.4.0"
drf-yasg = "^1.21.7"
isort = "^5.13.2"
black = "^24.8.0"
django-allauth = "^64.2.0"



[tool.poetry.dev-dependencies]
black = "^24.3.0"
flake8 = "^7.0.0"
ipdb = "^0.13.9"
ipython = "^8.2.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.2"
jupyterlab = "^4.0.11"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 100

[tool.isort]
profile = "black"
only_sections = "True"
no_lines_before = "FIRSTPARTY"
force_single_line = "True"
